requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
faker==22.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1

# ETL & Data Quality
great-expectations==0.18.8
prefect==2.14.10

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0
joblib==1.3.2

# Feature Engineering
category-encoders==2.6.3
feature-engine==1.6.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Configuration & Logging
python-dotenv==1.0.0
PyYAML==6.0.1
loguru==0.7.2

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.28.0
ipywidgets==8.1.1

# Power BI Connector (Python)
msal==1.26.0
